refactor to use get_secret_from_manager
